+++ libs/platform/dynamic-page/dynamic-page.component.ts
@@ -2,7 +2,6 @@ import {
     AfterContentInit,
     AfterViewInit,
     ChangeDetectionStrategy,
-    ChangeDetectorRef,
     Component,
     ContentChild,
     ContentChildren,
@@ -26,7 +25,7 @@ import {
 import { startWith } from 'rxjs/operators';
 
 import { NgTemplateOutlet } from '@angular/common';
-import { Nullable } from '@fundamental-ngx/cdk/utils';
+import { HasElementRef } from '@fundamental-ngx/cdk/utils';
 import { BreadcrumbComponent } from '@fundamental-ngx/core/breadcrumb';
 import {
     DynamicPageComponent as CoreDynamicPageComponent,
@@ -110,7 +109,7 @@ export class DynamicPageTabChangeEvent {
 })
 export class DynamicPageComponent
     extends BaseComponent
-    implements AfterContentInit, AfterViewInit, DoCheck, OnDestroy, PlatformDynamicPage
+    implements AfterContentInit, AfterViewInit, DoCheck, OnDestroy, PlatformDynamicPage, HasElementRef
 {
     /** Whether DynamicPage should snap on scroll */
     @Input() disableSnapOnScroll = false;
@@ -118,9 +117,6 @@ export class DynamicPageComponent
     @Input()
     @HostBinding('attr.role')
     role = 'region';
-
-    /** aria label for the page */
-    @Input() ariaLabel: Nullable<string>;
 
     /** Whether or not tabs should be stacked. */
     @Input() stackContent = false;
@@ -211,12 +207,7 @@ export class DynamicPageComponent
     protected _destroyRef = inject(DestroyRef);
 
     /** @hidden */
-    constructor(
-        protected _cd: ChangeDetectorRef,
-        public readonly elementRef: ElementRef<HTMLElement>
-    ) {
-        super(_cd);
-    }
+    readonly elementRef = inject<ElementRef<HTMLElement>>(ElementRef);
 
     /** toggle the visibility of the header on click of title area. */
     toggleCollapse(): void {
@@ -235,7 +226,7 @@ export class DynamicPageComponent
 
     /** @hidden */
     ngAfterViewInit(): void {
-        this._cd.detectChanges();
+        this.detectChanges();
 
         this._tabListComponent?.headerContainer.nativeElement.classList.add('fd-dynamic-page__tabs');
         this.collapsed = this._dynamicPageComponent.collapsed;
@@ -246,7 +237,7 @@ export class DynamicPageComponent
         /** Used to detect changes in projected components that displayed using templates,
          * https://github.com/angular/angular/issues/44112
          */
-        this._cd.markForCheck();
+        this.markForCheck();
     }
 
     /**